refactor(coap): move scopes into own module
